Add detail::value_init_when_moved_from class
This utility wraps a value and assigns it a value-initialized object
when the wrapper is moved from. This pattern is frequently used in
classes that need to set an integer/pointer to null when moved-from.
Using value_init_when_moved_from<int> as member gets rid of manually
defining move operations. It also lets us remove the swap()
overloads as they were only useful for implementing move operations.