Self-move assign is now a no-op
The switch to unique_ptr changed the behavior of self-move. Previously
the reservation was freed, now it's a no-op. This change, even though
observable, is not a breaking change as we follow the stdlib rule that
self-moved objects are still in a valid state which is true before and
after this change.